task_8/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
openai
python-dotenv

